server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
numpy==2.1.2
scipy==1.13.1
scikit-learn==1.5.2
livekit==1.0.18
livekit-api==1.0.7
sarvamai>=0.1.0
